requirements-api.txt
# API Server Dependencies
# Install with: pip install -r requirements-api.txt

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12  # For file uploads

# Data Validation
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0  # Async SQLite support

# Testing
hypothesis==6.115.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2  # For testing FastAPI endpoints
pytest-mock==3.14.0
coverage==7.6.1

# Configuration
pyyaml==6.0.2

# Utilities
python-dateutil==2.9.0
